changed ID type in migrations to bigInteger

+++ database/migrations/2018_12_23_124624_create_comments_table.php
@@ -15,9 +15,9 @@ class CreateCommentsTable extends Migration
     {
         Schema::enableForeignKeyConstraints();
         Schema::create('comments', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id')->unsigned();
-            $table->integer('photo_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->bigInteger('user_id')->unsigned();
+            $table->bigInteger('photo_id')->unsigned();
             $table->text('body');
             $table->timestamps();
